+++ __tests__/jsdom/components/SidebarChatList.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { render } from "@testing-library/react";
-import SidebarChatList from '@/components/Sidebar/ChatList/SidebarChatList';
+import SidebarChatList from '@/components/Sidebar/ChatList/component';
 import SidebarChatListItem from '@/components/Sidebar/ChatListItem/SidebarChatListItem';
 import { useState } from 'react'
 import { SidebarChatListItemProps } from '@/components/Sidebar/ChatListItem/interface';
